chore: Reorganize CLI commands for better readability
